Moved the global key/value args from advanced window into reduction settings window. This refs #224
